convert FullServersList into a dict
FullServersList was a list of lists and this meant that a lot of code
was not self-documenting and, thus, prone to error (you could get the
list index wrong for a bit of data, for eg).

Thus we create a new class (ServerXS) that describes the info we want
to store and make FullServersList a dictionary of that class.

This has allowed us to remove a lot of comments and code (particularily
in Sort()).

The Comment and Description variables were all made consistent as
Comment. They both pointed to the same data but were named differently
and the naming is now consistant with the 'comment' field in the
config file from which the data is sourced.