convert SshCommand/ConnectString to a list
Whilst security isn't likely to be an issue here it's, IMO, a good
idea to do things as safely as reasonably possible. In that light
commands and their args should be a list so as to clearly delineate
them with no possibility of confusion.

This was sort of at play here but, first, the command line was
constructed as a string. This was fixed.

ConnectString and SshCommand were renamed to CommandLine for
consistancy and because ConnectString stopped being a string. :)

Some of the command line was constructed in Connections.Load. I
believe that that, conceptually, should simple put the config
file together, usefully, for later usage. The construction of
the command line ought to be done in one place. This was also
done and the code simplified a little as False is the same as
an empty string or list when simply tested for truhtiness.